code/chatdoc-proxy/requirements.txt
python-multipart
pydantic==2.4.2
fastapi==0.110.0
uvicorn==0.29.0
pypeln==0.4.9
numpy==1.24.4
requests==2.31.0
pymilvus==2.4.3
bs4==0.0.1
lxml==5.2.2
jieba==0.42.1
tcvectordb==1.3.13
pdf2docx==0.5.8
PyYAML==6.0.1
elasticsearch==8.13.1
pymilvus==2.4.3
tcvectordb==1.3.13
tenacity==8.2.3
pillow==10.2.0
